Window is painted when resized
