+++ src/Consolonia.Core/Infrastructure/ConsoloniaPlatform.cs
@@ -41,6 +41,8 @@ namespace Consolonia.Core.Infrastructure
                 /*todo: need replacement? .Bind<IPlatformThreadingInterface>().ToSingleton<ConsoloniaPlatformThreadingInterface>()*/
                 .Bind<IRenderTimer>().ToConstant(new UiThreadRenderTimer(120))
                 .Bind<IDispatcherImpl>().ToConstant(new ManagedDispatcherImpl(null))
+                /*.Bind<IRenderTimer>().ToConstant(new SleepLoopRenderTimer(120))*/
+            /*SleepLoopRenderTimer : IRenderTimer*/
                 /*.Bind<IRenderLoop>().ToConstant(new RenderLoop()) todo: is internal now*/
                 .Bind<PlatformHotkeyConfiguration>().ToConstant(new PlatformHotkeyConfiguration(KeyModifiers.Control))
                 .Bind<IKeyboardDevice>().ToConstant(new ConsoleKeyboardDevice())
